Modified to remove MS compiler messages: ClpGubDynamicMatrix.cpp d:\coin\coinall-trunk\clp\src\clpgubdynamicmatrix.cpp(112) : warning C4244: '=' : conversion from 'const double' to 'float', possible loss of data d:\coin\coinall-trunk\clp\src\clpgubdynamicmatrix.cpp(115) : warning C4244: '=' : conversion from 'const double' to 'float', possible loss of data d:\coin\coinall-trunk\clp\src\clpgubdynamicmatrix.cpp(122) : warning C4244: '=' : conversion from 'const double' to 'float', possible loss of data d:\coin\coinall-trunk\clp\src\clpgubdynamicmatrix.cpp(129) : warning C4244: '=' : conversion from 'const double' to 'float', possible loss of data d:\coin\coinall-trunk\clp\src\clpgubdynamicmatrix.cpp(137) : warning C4244: '=' : conversion from 'const double' to 'float', possible loss of data d:\coin\coinall-trunk\clp\src\clpgubdynamicmatrix.cpp(139) : warning C4305: '=' : truncation from 'double' to 'float' d:\coin\coinall-trunk\clp\src\clpgubdynamicmatrix.cpp(144) : warning C4244: '=' : conversion from 'const double' to 'float', possible loss of data d:\coin\coinall-trunk\clp\src\clpgubdynamicmatrix.cpp(146) : warning C4305: '=' : truncation from 'double' to 'float'
I spoke to John about these warnings and attempted to follow his instructions.

+++ trunk/Clp/src/ClpGubDynamicMatrix.cpp
@@ -106,39 +106,39 @@ ClpGubDynamicMatrix::ClpGubDynamicMatrix(ClpSimplex * model, int numberSets,
   startColumn_ = ClpCopyOfArray(startColumn,numberGubColumns_+1);
   CoinBigIndex numberElements = startColumn_[numberGubColumns_];
   row_ = ClpCopyOfArray(row,numberElements);
-  element_ = new float[numberElements];
+  element_ = new double[numberElements];
   CoinBigIndex i;
   for (i=0;i<numberElements;i++)
     element_[i]=element[i];
-  cost_ = new float[numberGubColumns_];
+  cost_ = new double[numberGubColumns_];
   for (i=0;i<numberGubColumns_;i++) {
     cost_[i]=cost[i];
     // need sorted
     CoinSort_2(row_+startColumn_[i],row_+startColumn_[i+1],element_+startColumn_[i]);
   }
   if (lowerColumn) {
-    lowerColumn_ = new float[numberGubColumns_];
+    lowerColumn_ = new double[numberGubColumns_];
     for (i=0;i<numberGubColumns_;i++) 
       lowerColumn_[i]=lowerColumn[i];
   } else {
     lowerColumn_=NULL;
   }
   if (upperColumn) {
-    upperColumn_ = new float[numberGubColumns_];
+    upperColumn_ = new double[numberGubColumns_];
     for (i=0;i<numberGubColumns_;i++) 
       upperColumn_[i]=upperColumn[i];
   } else {
     upperColumn_=NULL;
   }
   if (upperColumn||lowerColumn) {
-    lowerSet_ = new float[numberSets_];
+    lowerSet_ = new double[numberSets_];
     for (i=0;i<numberSets_;i++) {
       if (lower[i]>-1.0e20)
 	lowerSet_[i]=lower[i];
       else
 	lowerSet_[i] = -1.0e30;
     }
-    upperSet_ = new float[numberSets_];
+    upperSet_ = new double[numberSets_];
     for (i=0;i<numberSets_;i++) {
       if (upper[i]<1.0e20)
 	upperSet_[i]=upper[i];
